kafka/requirements.txt
# confluent-kafka
kafka-python
mysql-connector-python
pyyaml